Update field type for Entry Hash

+++ ACHGenerator/DTO/ACHTransaction.cs
@@ -98,7 +98,7 @@ namespace ACHGenerator.DTO
         public int TransactionCode { get; set; }
 
         [ACHField(Position = 4, Length = 8)]
-        public int RDFIIdentification { get; set; }
+        public long RDFIIdentification { get; set; }
 
         [ACHField(Position = 12, Length = 1)]
         public int CheckDigit { get; set; }
@@ -137,7 +137,7 @@ namespace ACHGenerator.DTO
         public int EntryCount { get; set; }
 
         [ACHField(Position = 11, Length = 10)]
-        public int EntryHash { get; set; }
+        public long EntryHash { get; set; }
 
         [ACHField(Position = 21, Length = 12, Format = "0.00")]
         public decimal TotalDebitEntryDollarAmount { get; set; }
@@ -176,7 +176,7 @@ namespace ACHGenerator.DTO
         public int EntryCount { get; set; }
 
         [ACHField(Position = 22, Length = 10)]
-        public int EntryHash { get; set; }
+        public long EntryHash { get; set; }
 
         [ACHField(Position = 32, Length = 12, Format = "0.00")]
         public decimal TotalDebitEntryAmount { get; set; }
